refactor: consolidate IUpdateFileSystem and IProfileFileSystem into IFileSystem
Merge two independent file system abstractions into a single unified
IFileSystem interface in Common/, reducing duplication. The slim custom
interface exposes only the ~15 methods actually used by CliUpdater and
the profile sources.

- Add src/Func/Common/IFileSystem.cs (unified interface)
- Add src/Func/Common/PhysicalFileSystem.cs (production implementation)
- Remove IUpdateFileSystem, UpdateFileSystem, IProfileFileSystem, ProfileFileSystem
- Update all consumers and tests to use IFileSystem
- Rename RenameFile -> MoveFile, DeleteIfExistsAsync -> DeleteFileIfExistsAsync

Copilot-Session: d8c91ae0-c1cb-49ea-82e8-32537f3e5473
