+++ src/Func/Profiles/UserProfileSource.cs
@@ -1,6 +1,7 @@
 // Copyright (c) .NET Foundation. All rights reserved.
 // Licensed under the MIT License. See LICENSE in the project root for license information.
 
+using Azure.Functions.Cli.Common;
 using Azure.Functions.Cli.Configuration;
 
 namespace Azure.Functions.Cli.Profiles;
@@ -8,11 +9,11 @@ namespace Azure.Functions.Cli.Profiles;
 /// <summary>
 /// Loads user-level profiles from the Azure Functions CLI home directory.
 /// </summary>
-internal sealed class UserProfileSource(ProfileDocumentParser parser, IProfileFileSystem fileSystem, CliConfigurationPathsOptions configurationPaths)
+internal sealed class UserProfileSource(ProfileDocumentParser parser, IFileSystem fileSystem, CliConfigurationPathsOptions configurationPaths)
     : IProfileSource
 {
     private readonly ProfileDocumentParser _parser = parser ?? throw new ArgumentNullException(nameof(parser));
-    private readonly IProfileFileSystem _fileSystem = fileSystem ?? throw new ArgumentNullException(nameof(fileSystem));
+    private readonly IFileSystem _fileSystem = fileSystem ?? throw new ArgumentNullException(nameof(fileSystem));
     private readonly CliConfigurationPathsOptions _configurationPaths =
         configurationPaths ?? throw new ArgumentNullException(nameof(configurationPaths));
