--- src/Func/Update/UpdateFileSystem.cs
@@ -1,83 +1,0 @@
-// Copyright (c) .NET Foundation. All rights reserved.
-// Licensed under the MIT License. See LICENSE in the project root for license information.
-
-using System.IO.Compression;
-
-namespace Azure.Functions.Cli.Update;
-
-/// <inheritdoc cref="IUpdateFileSystem" />
-internal sealed class UpdateFileSystem : IUpdateFileSystem
-{
-    public string CreateTempDirectory()
-    {
-        string path = Path.Combine(Path.GetTempPath(), Path.GetRandomFileName());
-        Directory.CreateDirectory(path);
-        return path;
-    }
-
-    public string CreateTempDirectory(string siblingPath)
-    {
-        // siblingPath is expected to be the install directory itself. We create
-        // the temp dir alongside it (in the same parent) so it stays on the same
-        // volume without polluting the install directory.
-        string fullPath = Path.GetFullPath(siblingPath);
-        string parentDir = Directory.Exists(fullPath)
-            ? fullPath
-            : Path.GetDirectoryName(fullPath) ?? string.Empty;
-
-        if (string.IsNullOrEmpty(parentDir))
-        {
-            return CreateTempDirectory();
-        }
-
-        string path = Path.Combine(parentDir, ".func-update-" + Path.GetRandomFileName());
-        Directory.CreateDirectory(path);
-        return path;
-    }
-
-    public bool DirectoryExists(string path) => Directory.Exists(path);
-
-    public void MoveDirectory(string sourcePath, string destinationPath) =>
-        Directory.Move(sourcePath, destinationPath);
-
-    public void CopyDirectory(string sourcePath, string destinationPath)
-    {
-        DirectoryInfo source = new(sourcePath);
-        Directory.CreateDirectory(destinationPath);
-
-        foreach (FileInfo file in source.GetFiles())
-        {
-            file.CopyTo(Path.Combine(destinationPath, file.Name), overwrite: true);
-        }
-
-        foreach (DirectoryInfo subDir in source.GetDirectories())
-        {
-            CopyDirectory(subDir.FullName, Path.Combine(destinationPath, subDir.Name));
-        }
-    }
-
-    public void DeleteDirectory(string path) =>
-        Directory.Delete(path, recursive: true);
-
-    public IReadOnlyList<string> GetFiles(string directoryPath) =>
-        Directory.GetFiles(directoryPath, "*", SearchOption.AllDirectories);
-
-    public void RenameFile(string sourcePath, string destinationPath, bool overwrite = false) =>
-        File.Move(sourcePath, destinationPath, overwrite);
-
-    public void CopyFile(string sourcePath, string destinationPath) =>
-        File.Copy(sourcePath, destinationPath, overwrite: true);
-
-    public void DeleteFile(string path) => File.Delete(path);
-
-    public bool FileExists(string path) => File.Exists(path);
-
-    public async Task SaveStreamToFileAsync(string filePath, Stream content, CancellationToken cancellationToken)
-    {
-        await using FileStream file = File.Create(filePath);
-        await content.CopyToAsync(file, cancellationToken);
-    }
-
-    public void ExtractZip(string zipPath, string destinationDirectory) =>
-        ZipFile.ExtractToDirectory(zipPath, destinationDirectory);
-}
