+++ test/Func.Tests/Profiles/UserProfileSourceTests.cs
@@ -1,6 +1,7 @@
 // Copyright (c) .NET Foundation. All rights reserved.
 // Licensed under the MIT License. See LICENSE in the project root for license information.
 
+using Azure.Functions.Cli.Common;
 using Azure.Functions.Cli.Configuration;
 using Azure.Functions.Cli.Profiles;
 using NSubstitute;
@@ -15,7 +16,7 @@ public class UserProfileSourceTests
         string userHome = Path.Combine(Path.GetTempPath(), Guid.NewGuid().ToString("N"));
         string expectedPath = Path.Combine(userHome, CliConfigurationPathsOptions.ProfilesFileName);
         var paths = new CliConfigurationPathsOptions(userHome);
-        IProfileFileSystem fileSystem = Substitute.For<IProfileFileSystem>();
+        IFileSystem fileSystem = Substitute.For<IFileSystem>();
         fileSystem.ReadAllTextIfExistsAsync(expectedPath, CancellationToken.None).Returns(Task.FromResult<string?>(null));
         var source = new UserProfileSource(new ProfileDocumentParser(), fileSystem, paths);
         var context = new ProfileSourceContext(new DirectoryInfo(Path.GetTempPath()));
